remove break and replace with flag

+++ Final-Project/Install-7Zip.ps1
@@ -21,23 +21,27 @@ function Install-7Zip {
         $Flag = "7zip_installed"
         if (Test-Path "C:\$Flag") {
             Write-Host "7Zip is already installed" -ForegroundColor Yellow
-            break
+            $Skip = $true
         }
     }
     
     process {
-        $dlurl = 'https://7-zip.org/' + (Invoke-WebRequest -UseBasicParsing -Uri 'https://7-zip.org/' | Select-Object -ExpandProperty Links | Where-Object { ($_.outerHTML -match 'Download') -and ($_.href -like "a/*") -and ($_.href -like "*-x64.exe") } | Select-Object -First 1 | Select-Object -ExpandProperty href)
-        # modified to work without IE
-        # above code from: https://perplexity.nl/windows-powershell/installing-or-updating-7-zip-using-powershell/
-        $installerPath = Join-Path $env:TEMP (Split-Path $dlurl -Leaf)
-        Invoke-WebRequest $dlurl -OutFile $installerPath
-        Start-Process -FilePath $installerPath -Args "/S" -Verb RunAs -Wait
-        Remove-Item $installerPath
+        if ($Skip -ne $true) {
+            $dlurl = 'https://7-zip.org/' + (Invoke-WebRequest -UseBasicParsing -Uri 'https://7-zip.org/' | Select-Object -ExpandProperty Links | Where-Object { ($_.outerHTML -match 'Download') -and ($_.href -like "a/*") -and ($_.href -like "*-x64.exe") } | Select-Object -First 1 | Select-Object -ExpandProperty href)
+            # modified to work without IE
+            # above code from: https://perplexity.nl/windows-powershell/installing-or-updating-7-zip-using-powershell/
+            $installerPath = Join-Path $env:TEMP (Split-Path $dlurl -Leaf)
+            Invoke-WebRequest $dlurl -OutFile $installerPath
+            Start-Process -FilePath $installerPath -Args "/S" -Verb RunAs -Wait
+            Remove-Item $installerPath
 
-        New-Item -Path "C:\" -Name $Flag -ItemType File
+            New-Item -Path "C:\" -Name $Flag -ItemType File
+        }
     }
     
     end {
-        Write-Host "7Zip installed successfully" -ForegroundColor Green
+        if ($Skip -ne $true) {
+            Write-Host "7Zip installed successfully" -ForegroundColor Green
+        }
     }
 }#Install-7Zip
